+++ src/test/java/picocli/AtFileTest.java
@@ -776,39 +776,4 @@ public class AtFileTest {
                 "  -V, --version          Print version information and exit.%n");
         assertEquals(expected, actual);
     }
-
-    @Ignore("Requires #985 Show end-of-options `--` in usage help options list")
-    @Test
-    public void testEndOfOptionsDelimiter() {
-        @Command(name = "A", mixinStandardHelpOptions = true,
-                //FIXME showEndOfOptionsDelimiterInUsageHelp = true,
-                description = "... description ...")
-        class A {
-            @Parameters(arity = "1", description = "The file.")
-            private File file;
-
-            @Option(names = {"-x", "--long"}, arity="0..*", description = "Option with multiple params.")
-            private String params;
-        }
-
-        String actual = new CommandLine(new A()).getUsageMessage();
-        String expected = String.format("" +
-                "Usage: A [-hV] [-x[=<params>...]] [--] <file>%n" +
-                "... description ...%n" +
-                "      <file>             The file.%n" +
-                "  --                     End of options. All following arguments are positional%n" +
-                "                           parameters.%n" +
-                "  -x=[<params>...]       Option with multiple params.%n");
-        assertEquals(expected, actual);
-    }
-
-@Test
-public void test() {
-    String txt = "?String" + '\uD800';
-    System.out.println(txt); // prints ?String?
-    for (byte b : txt.getBytes()) {
-        System.out.print(" 0x" + Integer.toHexString(b));
-    } // gives  0x3f 0x53 0x74 0x72 0x69 0x6e 0x67 0x3f
-    System.out.println();
-}
 }